refactor: move bash tools to tools/ directory
Consolidate all tool implementations under tools/ directory:

**Changes:**
- Move src/bash-tools.ts → src/tools/bash.ts
- Update import paths:
  - src/index.ts: './bash-tools.ts' → './tools/bash.ts'
  - test/session.test.ts: '../src/bash-tools.ts' → '../src/tools/bash.ts'
- Fix relative import in bash.ts: './types.ts' → '../types.ts'
- Update src/tools/index.ts with bash tools exports and documentation

**Final Structure:**
```
src/tools/
  ├── index.ts      # Tool exports & documentation
  ├── skills.ts     # Skills tool
  └── bash.ts       # Bash tools (bash, readFile, writeFile)

src/skills/
  ├── provider.ts   # Skills provider
  ├── importer.ts   # Git importer
  ├── import-spec.ts # Import spec parser
  └── index.ts      # Skills exports
```

**Benefits:**
- All tool implementations in one place
- Clear separation: tools/ for implementations, skills/ for domain logic
- Consistent structure ready for future tools (git, todo, etc.)

All 143 tests passing ✓
Build successful ✓
