+++ packages/agent-worker/test/session.test.ts
@@ -612,7 +612,7 @@ describe('createTools edge cases', () => {
 
 describe('bash-tools integration', () => {
   test('createBashTools returns tools and toolkit', async () => {
-    const { createBashTools } = await import('../src/bash-tools.ts')
+    const { createBashTools } = await import('../src/tools/bash.ts')
 
     const { tools, toolkit } = await createBashTools({
       files: { 'test.txt': 'hello world' },
@@ -625,7 +625,7 @@ describe('bash-tools integration', () => {
   })
 
   test('createBashTools respects includeReadFile option', async () => {
-    const { createBashTools } = await import('../src/bash-tools.ts')
+    const { createBashTools } = await import('../src/tools/bash.ts')
 
     const { tools } = await createBashTools({
       files: {},
@@ -636,7 +636,7 @@ describe('bash-tools integration', () => {
   })
 
   test('createBashTools respects includeWriteFile option', async () => {
-    const { createBashTools } = await import('../src/bash-tools.ts')
+    const { createBashTools } = await import('../src/tools/bash.ts')
 
     const { tools } = await createBashTools({
       files: {},
@@ -647,7 +647,7 @@ describe('bash-tools integration', () => {
   })
 
   test('bash tool executes commands', async () => {
-    const { createBashTools } = await import('../src/bash-tools.ts')
+    const { createBashTools } = await import('../src/tools/bash.ts')
 
     const { tools } = await createBashTools({
       files: { 'test.txt': 'hello world' },
@@ -665,7 +665,7 @@ describe('bash-tools integration', () => {
   })
 
   test('readFile tool reads files', async () => {
-    const { createBashTools } = await import('../src/bash-tools.ts')
+    const { createBashTools } = await import('../src/tools/bash.ts')
 
     const { tools } = await createBashTools({
       files: { 'hello.txt': 'Hello, World!' },
@@ -678,7 +678,7 @@ describe('bash-tools integration', () => {
   })
 
   test('writeFile tool writes files', async () => {
-    const { createBashTools } = await import('../src/bash-tools.ts')
+    const { createBashTools } = await import('../src/tools/bash.ts')
 
     const { tools } = await createBashTools({ files: {} })
 
@@ -692,7 +692,7 @@ describe('bash-tools integration', () => {
   })
 
   test('createBashToolsFromFiles helper', async () => {
-    const { createBashToolsFromFiles } = await import('../src/bash-tools.ts')
+    const { createBashToolsFromFiles } = await import('../src/tools/bash.ts')
 
     const { tools } = await createBashToolsFromFiles({
       'src/index.ts': 'console.log("hello")',
@@ -707,7 +707,7 @@ describe('bash-tools integration', () => {
   })
 
   test('tools have correct parameter schemas', async () => {
-    const { createBashTools } = await import('../src/bash-tools.ts')
+    const { createBashTools } = await import('../src/tools/bash.ts')
 
     const { tools } = await createBashTools({ files: {} })
 
@@ -727,7 +727,7 @@ describe('bash-tools integration', () => {
   })
 
   test('tools work with AgentSession', async () => {
-    const { createBashTools } = await import('../src/bash-tools.ts')
+    const { createBashTools } = await import('../src/tools/bash.ts')
 
     const { tools } = await createBashTools({
       files: { 'data.json': '{"key": "value"}' },
